refactor: DispatcherServlet을 mvc 모듈 하위로 이동
