+++ mvc/src/main/java/com/interface21/webmvc/servlet/NotFoundAdapterException.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
 public class NotFoundAdapterException extends RuntimeException {
